fix(exoplayer): handle multiple discontinuity events
Send pending updates before adding discontinuity for error

When handling a playback error that originates from a future item in
the playlist, we added support for jumping to that item first,
ensuring the errors 'happen' for the right 'current item'.
See 79b688e.

However, when we add this new position discontinuity to the
playback state, there may already be other position discontinuities
pending from other parts of the code that executed before the
error. As we can't control that in this case (because it's part
of a generic try/catch block), we need to send any pending
updates first before handling the new change.

Issue: #1483
#cherrypick
Refs: https://github.com/androidx/media/commit/727645179b3c26e495540f4445216fd035cc7654
PiperOrigin-RevId: 646968309